stylized application

+++ src/app/pages/products/products/product-list/product-list.component.ts
@@ -3,6 +3,7 @@ import { Product } from '../../../../core/interfaces/product';
 import { ProductService } from '../../../../services/product.service';
 import { firstValueFrom } from 'rxjs';
 import { NotificationService } from '../../../../services/notification.service';
+import { ModalService } from '../../../../services/modal.service';
 
 @Component({
   selector: 'app-product-list',
@@ -17,7 +18,11 @@ export class ProductListComponent implements OnInit {
   public search: string = '';
   public size: number = 5;
 
-  constructor(private _productService: ProductService, private _notificationService: NotificationService) {}
+  constructor(
+    private _productService: ProductService,
+    private _notificationService: NotificationService,
+    private _modalService: ModalService
+  ) {}
   ngOnInit(): void {
     this.loadProducts();
   }
@@ -28,28 +33,47 @@ export class ProductListComponent implements OnInit {
       let resp: Product[] = await firstValueFrom(
         this._productService.getProducts()
       );
-      console.log(this.products);
       this._products = resp;
       this.products = this._products.slice(0, this.size);
       this.loader = false;
-    } catch (error) {
+    } catch (error: any) {
       this.loader = false;
       console.log(error);
+      this._modalService.openModal(
+        'error',
+        'Error',
+        error?.error?.message || error?.message || JSON.parse(error)
+      );
     }
   }
 
   public async deleteProduct(id: string | undefined) {
     try {
-      let resp = await firstValueFrom(
-        this._productService.deleteProduct(id as string)
+      this._modalService.openModal(
+        'error',
+        'Delete product id:' + id,
+        'Are you sure you want to delete this product?',
+        true
       );
-      console.log('deleteProduct', resp);
-      this._notificationService.showSuccess('Product deleted');
-      this.products = this.products.filter(
-        (product: Product) => product.id !== id
-      );
-    } catch (error) {
+
+      this._modalService.eventOnOk.subscribe(async (isOk: boolean) => {
+        if (isOk) {
+          let resp = await firstValueFrom(
+            this._productService.deleteProduct(id as string)
+          );
+          this._notificationService.showSuccess('Product deleted');
+          this.products = this.products.filter(
+            (product: Product) => product.id !== id
+          );
+        }
+      });
+    } catch (error: any) {
       console.log(error);
+      this._modalService.openModal(
+        'error',
+        'Error',
+        error?.error?.message || error?.message || JSON.parse(error)
+      );
     }
   }
 
@@ -73,5 +97,4 @@ export class ProductListComponent implements OnInit {
         product?.description?.toLowerCase().includes(this.search.toLowerCase())
     );
   }
-
 }
